fix: move analytics components to TEI directory
