+++ src/components/analytics/SelectProgramStage.js
@@ -2,7 +2,7 @@ import React from 'react'
 import i18n from '@dhis2/d2-i18n'
 import PropTypes from '@dhis2/prop-types'
 import { SelectField } from './SelectField'
-import { createInitialValues } from '../../pages/Analytics/helper'
+import { createInitialValues } from '../../pages/Analytics/TEI/helper'
 import { isValidValue } from '../../utils/validators'
 
 export const SelectProgramStage = ({
