+++ src/pages/Analytics/TEI/helper.js
@@ -1,7 +1,7 @@
 import omit from 'lodash/omit'
 import map from 'lodash/map'
-import { validateObjectByProperty } from '../../utils/validators'
-import { findProgramNameById } from '../../utils/utils'
+import { validateObjectByProperty } from '../../../utils/validators'
+import { findProgramNameById } from '../../../utils/utils'
 
 export const WHO_NUTRITION = 'WHO_NUTRITION'
